refactor(workout-plan): extract media integration (WP3.4h)
